Repository with tools and animations
I have made a batch manifest creator and around 60 animations. Around 35 of those are unique and the others are inversed colors so that everyone could choose an orange or black background. I need to add the animations and previews in the repository, I just finished working on the animations, ill make some more and add everything!